+++ api/auth/jwtMiddleware.js
@@ -1,10 +1,10 @@
 const { jwtExtractor, verifyWebToken } = require('./jwtUtils');
-const { deserializeUser } = require('./serialization');
+const { getUserByID: gu } = require('../db');
 const logger = require('../logger')('jwt middleware');
 
 /**
  * Extracts and verifies the JWT in the request Authorization header.
- * Attaches the JWT payload and user to the request, if JWT is valid.
+ * Attaches the user to the request, if JWT is valid.
  * 401/403 responses are handled by `can` and `loggedIn`.
  * @name jwtMiddleware
  * @function
@@ -17,24 +17,28 @@ const jwtMiddleware = async (
   res,
   next,
   {
-    deserialize = deserializeUser,
+    getUserByID = gu,
     extractor = jwtExtractor,
     verifyToken = verifyWebToken
   } = {}
 ) => {
   const jwt = extractor(req);
-  const payload = jwt ? verifyToken(jwt) : false;
-
-  if (!payload) return next();
-
   try {
-    const user = await deserialize(payload.sub);
+    const claims = jwt ? await verifyToken(jwt) : false;
+    if (!claims) return next();
+
+    // some values like group and application profile variables
+    // are returned by the claims for conviences, but not retrieved
+    // by the standard getUser call to Okta, so these values should
+    // be passed in as additional values when possible.
+    const { uid, ...additionalValues } = claims;
+    const user = await getUserByID(uid, { additionalValues });
+    // const user = claims;
     if (user) {
       req.user = user;
-      req.payload = payload;
     }
   } catch (err) {
-    logger.error(err);
+    logger.error(`error message: ${err.message}`);
   }
   return next();
 };