+++ api/endpoint-tests/utils.js
@@ -3,10 +3,12 @@ const FormData = require('form-data');
 const knex = require('knex');
 const knexConfig = require('../knexfile');
 
-const baseURL = () =>
-  `http://${process.env.API_HOST || 'localhost'}:${process.env.API_PORT ||
-    process.env.PORT ||
-    8000}`;
+const { API_HOST, API_PORT, PORT } = process.env;
+
+const host = API_HOST || 'localhost';
+const port = API_PORT || PORT || 8000;
+
+const baseURL = () => `http://${host}:${port}`;
 
 const axiosDefaults = {
   baseURL: baseURL(),
@@ -15,45 +17,29 @@ const axiosDefaults = {
 
 const api = axios.create(axiosDefaults);
 
-const login = (
-  username = 'all-permissions-and-state',
-  password = 'password'
-) => {
-  return api
-    .post('/auth/login/nonce', { username })
-    .then(res => res.data.nonce)
-    .then(nonce => api.post('/auth/login', { username: nonce, password }))
-    .then(res => {
-      if (res.status !== 200) return new Error('Failed to login', res);
-      return res;
-    })
-    .then(res => {
-      const options = {
-        ...axiosDefaults,
-        headers: {
-          Authorization: `Bearer ${res.data.token}`
-        }
-      };
-      return axios.create(options);
-    })
-    .catch(error => error);
+const login = token => {
+  const jwt = token || 'all-permissions'
+  const options = {
+    ...axiosDefaults,
+    headers: {
+      Authorization: `Bearer ${jwt}`
+    }
+  };
+  return axios.create(options);
 };
 
-const unauthenticatedTest = (method, url) =>
+const unauthenticatedTest = (method, url) => {
   it('when unauthenticated', async () => {
     const response = await api[method](url);
     expect(response.status).toEqual(401);
     expect(response.data).toBeFalsy();
   });
+};
 
 const unauthorizedTest = (method, url) => {
   it('when unauthorized', async () => {
-    // this user has no permissions
-    const response = await login(
-      'no-permissions',
-      'password'
-    ).then(authenticatedClient => authenticatedClient[method](url));
-
+    const authenticatedClient = login('no-permissions');
+    const response = await authenticatedClient[method](url);
     expect(response.status).toEqual(403);
     expect(response.data).toBeFalsy();
   });