+++ api/routes/auth/activities/get.endpoint.js
@@ -11,7 +11,8 @@ describe('auth activities endpoint | GET /auth/activities', () => {
   unauthorizedTest('get', url);
 
   it('when authenticated', async () => {
-    const response = await login().then(api => api.get(url));
+    const api = login();
+    const response = await api.get(url);
 
     expect(response.status).toEqual(200);
     expect(response.data).toMatchSnapshot();