+++ api/seeds/test/roles.js
@@ -1,121 +1,61 @@
+// 10xx IDs refer to activities
+// 11xx IDs refer to roles
+
+// For testing, put in our hard-coded stuff so we know what IDs to expect.
+const activities = [
+  { id: 1001, name: 'view-users' },
+  { id: 1002, name: 'view-roles' },
+  { id: 1003, name: 'submit-federal-response' },
+  { id: 1004, name: 'submit-clearance' },
+  { id: 1005, name: 'edit-comments' },
+  { id: 1006, name: 'submit-document' },
+  { id: 1007, name: 'submit-state-response' },
+  { id: 1008, name: 'create-draft' },
+  { id: 1009, name: 'edit-document' },
+  { id: 1010, name: 'edit-response' },
+  { id: 1011, name: 'view-document' },
+  { id: 1012, name: 'view-affiliations' },
+  { id: 1013, name: 'edit-affiliations' },
+  { id: 1014, name: 'approve-state-access' },
+];
+exports.activities = activities;
+
+const roles = [
+  { id: 1101, isActive: true, name: 'eAPD Admin' },
+  { id: 1102, isActive: false, name: 'eAPD Federal Analyst' },
+  { id: 1103, isActive: false, name: 'eAPD Federal Leadership' },
+  { id: 1104, isActive: false, name: 'eAPD Federal SME' },
+  { id: 1105, isActive: true, name: 'eAPD State Coordinator' },
+  { id: 1106, isActive: false, name: 'eAPD State SME' },
+  { id: 1107, isActive: true, name: 'eAPD State Admin' },
+];
+exports.roles = roles;
+
+// grant 'eAPD Admin' role all activities
+const adminRole = roles.find(role => role.name === 'eAPD Admin');
+const adminActivities = activities.map(activity => ({
+  role_id: adminRole.id,
+  activity_id: activity.id
+}));
+
+// 'eAPD State Coordinator'
+const stateAdminRole = roles.find(role => role.name === 'eAPD State Admin');
+const stateAdminActivities = [
+  'view-document',
+  'submit-document',
+  'submit-state-response',
+  'create-draft',
+  'edit-document',
+  'edit-response',
+  'approve-state-access'
+].map(activityName => ({
+  role_id: stateAdminRole.id,
+  activity_id: activities.find(activity => activity.name === activityName).id
+}));
+
 exports.seed = async knex => {
-  // 10xx IDs refer to activities
-  // 11xx IDs refer to roles
-
-  // For testing, put in our hard-coded stuff so we know what IDs to expect.
-  await knex('auth_activities').insert({ id: 1001, name: 'view-users' });
-  await knex('auth_activities').insert({ id: 1002, name: 'add-users' });
-  await knex('auth_activities').insert({ id: 1003, name: 'view-roles' });
-  await knex('auth_activities').insert({ id: 1004, name: 'create-roles' });
-  await knex('auth_activities').insert({ id: 1005, name: 'edit-roles' });
-  await knex('auth_activities').insert({
-    id: 1006,
-    name: 'submit-federal-response'
-  });
-  await knex('auth_activities').insert({ id: 1007, name: 'submit-clearance' });
-  await knex('auth_activities').insert({ id: 1008, name: 'edit-comments' });
-  await knex('auth_activities').insert({ id: 1009, name: 'submit-document' });
-  await knex('auth_activities').insert({
-    id: 1010,
-    name: 'submit-state-response'
-  });
-  await knex('auth_activities').insert({ id: 1011, name: 'create-draft' });
-  await knex('auth_activities').insert({ id: 1012, name: 'edit-document' });
-  await knex('auth_activities').insert({ id: 1013, name: 'edit-response' });
-  await knex('auth_activities').insert({ id: 1014, name: 'delete-users' });
-  await knex('auth_activities').insert({ id: 1015, name: 'view-document' });
-
-  await knex('auth_roles').insert({ id: 1101, name: 'admin', isActive: true });
-  await knex('auth_roles').insert({
-    id: 1102,
-    name: 'federal analyst',
-    isActive: false
-  });
-  await knex('auth_roles').insert({
-    id: 1103,
-    name: 'federal leadership',
-    isActive: false
-  });
-  await knex('auth_roles').insert({
-    id: 1104,
-    name: 'federal SME',
-    isActive: false
-  });
-  await knex('auth_roles').insert({
-    id: 1105,
-    name: 'state coordinator',
-    isActive: true
-  });
-  await knex('auth_roles').insert({
-    id: 1106,
-    name: 'state SME',
-    isActive: false
-  });
-
-  // admin
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1001
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1002
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1003
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1004
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1005
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1006
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1007
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1008
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1009
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1010
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1011
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1012
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1013
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1014
-  });
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1101,
-    activity_id: 1015
-  });
-
-  await knex('auth_role_activity_mapping').insert({
-    role_id: 1102,
-    activity_id: 1005
-  });
+  await knex('auth_activities').insert(activities);
+  await knex('auth_roles').insert(roles);
+  await knex('auth_role_activity_mapping').insert(adminActivities);
+  await knex('auth_role_activity_mapping').insert(stateAdminActivities);
 };